Zaturanva/Zaturanva.Common/Pieces/Pawn.cs
using LanguageExt;

using Zaturanva.Common.ChessBoard;
using Zaturanva.Common.Colors;
using Zaturanva.Common.Contestants.PlayerManagement;
using Zaturanva.Common.Games;

namespace Zaturanva.Common.Pieces;

public class Pawn : IPiece
{
	public Color Color { get; set; }

	public required IPlayer Owner { get; set; }

	public Option<Color> CapturedBy { get; set; } = Option<Color>.None;

	public required Option<Coordinates> Location { get; set; }

	public bool CanMoveTo(GameState game, Coordinates destination)
		=> Location.Match(
			currentLocation
				=> IsMovementAllowed(game.Board, currentLocation, destination)
				   && game.IsMoveAllowedByStandardRules(this, destination),
			() => false
		);

	private static bool IsMovementAllowed(
		Board board,
		Coordinates currentLocation,
		Coordinates destination
	)
	{
		int xDifference = Math.Abs(destination.X - currentLocation.X);
		int yDifference = destination.Y - currentLocation.Y;
		bool isOneStepForward = (xDifference == 0)
								&& (yDifference == 1)
								&& board[destination]
									.IsVacant();
		bool isDiagonalCapture = (xDifference == 1)
								 && (yDifference == 1)
								 && board[destination]
									 .IsOccupied();
		return isOneStepForward || isDiagonalCapture;
	}
}